Refactor test files to account for new naming standard of files
